chore: update database query permission checks
