+++ frontend/src/utils/index.ts
@@ -17,7 +17,6 @@ export * from "./deployment";
 export * from "./string";
 export * from "./tab";
 export * from "./task";
-export * from "./policy";
 export * from "./sheet";
 export * from "./pev2";
 export * from "./web-storage";
